更新了实体类，使其对应更新后的数据库表

+++ src/main/java/club/boyuan/official/teammatching/controller/CommunityController.java
@@ -28,7 +28,7 @@ public class CommunityController {
     public PostDetailResponse<PostDetailResponse.PostCreateVO> createPost(@Valid @RequestBody CreatePostRequest request) {
         // 模拟从 JWT 中获取当前登录用户 ID（暂时写死，之后从 SecurityContext 获取）
         Integer currentUserId = 34343333;
-        Long newPostId = communityService.createNewPost(request, currentUserId);
+        Long newPostId = communityService.createNewPost(request, currentUserId).longValue();
         PostDetailResponse.PostCreateVO vo = new PostDetailResponse.PostCreateVO(newPostId, "发布成功，待审核");
         // 调用 Service 执行发布逻辑
         return PostDetailResponse.success(vo);
